updated simulation time predictions
