+++ delphin_6_automation/database_interactions/db_templates/delphin_entry.py
@@ -26,6 +26,7 @@ class Delphin(mongoengine.Document):
     simulated = mongoengine.DateTimeField()
     simulating = mongoengine.BooleanField(default=False)
     simulation_time = mongoengine.FloatField()
+    estimated_simulation_time = mongoengine.IntField()
     queue_priority = mongoengine.IntField(default=1)
     sample_data = mongoengine.DictField()
     restart_data = mongoengine.DictField()
