--- delphin_6_automation/database_interactions/logging_interactions.py
@@ -1,24 +1,0 @@
-__author__ = "Christian Kongsgaard"
-__license__ = 'MIT'
-
-# -------------------------------------------------------------------------------------------------------------------- #
-# IMPORTS
-
-# Modules
-
-# RiBuild Modules
-from delphin_6_automation.logging.ribuild_logger import ribuild_logger
-
-# Logger
-logger = ribuild_logger()
-
-# -------------------------------------------------------------------------------------------------------------------- #
-# RIBuild
-
-
-def connect_to_log_database():
-    pass
-
-
-def upload_logs():
-    pass
