+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy==1.15
+numpy
 pandas
 scipy
 matplotlib
